moved cseditor components into tools to avoid code duplicatoin
