+++ src/app/widgets/cseditor/components/attacks.h
@@ -2,13 +2,13 @@
 #define CSEDITOR_ATTACKS_H
 
 #include "CS/attacks.h"
-#include "component.h"
+#include "tools/component.h"
 #include <QSpinBox>
 #include <QList>
 
 namespace CSEditor {
 
-class Attacks : public Component
+class Attacks : public Tools::Component
 {
     Q_OBJECT
 public:
