+++ src/app/widgets/cseditor/components/equipment.h
@@ -1,7 +1,7 @@
 #ifndef CSEDITOR_EQUIPMENT_H
 #define CSEDITOR_EQUIPMENT_H
 
-#include "component.h"
+#include "tools/component.h"
 #include "CS/equipment.h"
 #include <QPushButton>
 #include <QList>
@@ -10,7 +10,7 @@ namespace CSEditor {
 
 class Item;
 
-class Equipment : public Component
+class Equipment : public Tools::Component
 {
     Q_OBJECT
 public:
