+++ src/app/widgets/gamecs/components/basicstat.h
@@ -2,14 +2,14 @@
 #define GAMECS_BASICSTAT_H
 
 #include "CS/basicstat.h"
-#include "component.h"
+#include "tools/component.h"
 
 #include <QWidget>
 #include <QBoxLayout>
 
 namespace GameCS {
 
-class BasicStat : public Component
+class BasicStat : public Tools::Component
 {
     Q_OBJECT
 public:
