+++ src/app/widgets/gamecs/components/descriptor.cpp
@@ -8,7 +8,7 @@ namespace GameCS {
 /******************************************************************************/
 
 Descriptor::Descriptor(CS::Descriptor *descriptor, QWidget *parent)
-    : Component(descriptor->getTitle(), parent), descriptor(descriptor) {
+    : Tools::Component(descriptor->getTitle(), parent), descriptor(descriptor) {
     bodyAdd(new QLabel(descriptor->getDescription()));
     setStyleSheet("QLabel { font-size: 18px; }"
                   "QPushButton { font-size: 14px; border: 1px solid #282828; "
